Consolidate short files: merge format components, lint rule-helpers, and performance-rule-ids
- Merge default-component-instances.ts + format-components.ts into
  default-format-components.ts (3 files → 1)
- Merge rule-helpers.ts into rule-base-helpers.ts (2 files → 1)
- Merge performance-rule-ids.ts into rule-level-presets.ts (2 files → 1)
- Update all imports and tests to point to new locations
- Net reduction: 4 source files removed

Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/723d42b1-bdb8-4c12-a785-b28c8b072065
