+++ src/format/src/components/default-format-components.ts
@@ -1,27 +1,54 @@
 import type { GameMakerAstNode } from "@gmloop/core";
 
-import { defaultGmlFormatComponentImplementations } from "./default-component-instances.js";
-import type { GmlFormatComponentBundle } from "./format-types.js";
+import { handleComments, printComment } from "../comments/index.js";
+import { LogicalOperatorsStyle } from "../options/logical-operators-style.js";
+import { gmlParserAdapter } from "../parsers/index.js";
+import { print } from "../printer/index.js";
+import { normalizeGmlFormatComponents } from "./format-component-normalizer.js";
+import type { GmlFormatComponentBundle, GmlFormatComponentContract } from "./format-types.js";
+
+/**
+ * Default implementation bundle wiring the canonical parser, printer, and
+ * comment handlers. This is the single point where concrete adapters are
+ * assembled into the component contract.
+ */
+export const defaultGmlFormatComponentImplementations: GmlFormatComponentContract = Object.freeze({
+    gmlParserAdapter,
+    print,
+    handleComments,
+    printComment,
+    LogicalOperatorsStyle
+});
+
+/**
+ * The immutable, normalized format component bundle used by the GML Prettier plugin.
+ * This constant is initialized once at module load time and never changes.
+ *
+ * Components include:
+ * - Parsers for converting GML source to AST
+ * - Printers for converting AST back to formatted GML
+ * - Format options and their defaults
+ */
+export const gmlFormatComponents: GmlFormatComponentBundle = Object.freeze(
+    normalizeGmlFormatComponents(createDefaultGmlFormatComponents())
+);
 
 export function createDefaultGmlFormatComponents(): GmlFormatComponentBundle {
-    const { gmlParserAdapter, print, handleComments, printComment, LogicalOperatorsStyle } =
-        defaultGmlFormatComponentImplementations;
-
     return {
         parsers: {
-            "gml-parse": gmlParserAdapter,
-            gmlParserAdapter
+            "gml-parse": defaultGmlFormatComponentImplementations.gmlParserAdapter,
+            gmlParserAdapter: defaultGmlFormatComponentImplementations.gmlParserAdapter
         },
         printers: {
             "gml-ast": {
-                print,
+                print: defaultGmlFormatComponentImplementations.print,
                 // Accept any for the runtime types coming from the AST and comment
                 // helpers, satisfying TypeScript without adding deep imports.
                 isBlockComment: (comment: GameMakerAstNode) => comment?.type === "CommentBlock",
                 canAttachComment: (node: GameMakerAstNode) =>
                     node?.type && !node.type.includes("Comment") && node?.type !== "EmptyStatement",
-                printComment,
-                handleComments
+                printComment: defaultGmlFormatComponentImplementations.printComment,
+                handleComments: defaultGmlFormatComponentImplementations.handleComments
             }
         },
         options: {
