+++ src/format/src/components/index.ts
@@ -1,3 +1,2 @@
-export * from "./default-component-instances.js";
 export * from "./default-format-components.js";
 export * from "./format-component-normalizer.js";
