--- src/lint/src/configs/performance-rule-ids.ts
@@ -1,7 +1,0 @@
-export const PERFORMANCE_OVERRIDE_RULE_IDS = Object.freeze([
-    "gml/prefer-hoistable-loop-accessors",
-    "gml/prefer-loop-invariant-expressions",
-    "gml/prefer-struct-literal-assignments",
-    "gml/no-globalvar",
-    "gml/prefer-string-interpolation"
-]);
