--- src/lint/src/rules/gml/rule-helpers.ts
@@ -1,19 +1,0 @@
-import type { Rule } from "eslint";
-
-export function readObjectOption(context: Rule.RuleContext): Record<string, unknown> {
-    if (!Array.isArray(context.options)) {
-        return Object.freeze({});
-    }
-
-    const [rawOption] = context.options;
-    if (!rawOption || typeof rawOption !== "object") {
-        return Object.freeze({});
-    }
-
-    return rawOption as Record<string, unknown>;
-}
-
-export function shouldReportUnsafe(context: Rule.RuleContext): boolean {
-    const option = readObjectOption(context).reportUnsafe;
-    return option === undefined ? true : option === true;
-}
