+++ src/lint/src/rules/gml/rules/prefer-hoistable-loop-accessors-rule.ts
@@ -6,12 +6,13 @@ import {
     createMeta,
     getNodeStartIndex,
     isAstNodeWithType,
+    readObjectOption,
     resolveLocFromIndex,
+    shouldReportUnsafe,
     walkAstNodes,
     walkAstNodesWithParent
 } from "../rule-base-helpers.js";
 import type { GmlRuleDefinition } from "../rule-definition.js";
-import { readObjectOption, shouldReportUnsafe } from "../rule-helpers.js";
 
 const DEFAULT_HOIST_ACCESSORS = Object.freeze({
     array_length: "len"
