+++ src/lint/src/services/index.ts
@@ -1,4 +1,4 @@
-import { PERFORMANCE_OVERRIDE_RULE_IDS } from "../configs/performance-rule-ids.js";
+import { PERFORMANCE_OVERRIDE_RULE_IDS } from "../configs/rule-level-presets.js";
 import { featherManifest } from "../rules/feather/manifest.js";
 
 export const services = Object.freeze({
